+++ lib/protobuf/nats.rb
@@ -6,11 +6,14 @@ require "protobuf/rpc/service_directory"
 
 require "nats/io/client"
 
-require "protobuf/nats/errors"
+require "protobuf/nats/response_muxer"
+require "protobuf/nats/super_subscription_manager"
+
 require "protobuf/nats/client"
-require "protobuf/nats/server"
-require "protobuf/nats/runner"
 require "protobuf/nats/config"
+require "protobuf/nats/errors"
+require "protobuf/nats/runner"
+require "protobuf/nats/server"
 
 module Protobuf
   module Nats
